+++ src/jwk.rs
@@ -2,11 +2,14 @@
 //! This crate contains types only for working JWK and JWK Sets
 //! This is only meant to be used to deal with public JWK, not generate ones.
 //! Most of the code in this file is taken from https://github.com/lawliet89/biscuit but
-// tweaked to remove the private bits as it's not the goal for this crate currently.
-//!
-use crate::Algorithm;
+//! tweaked to remove the private bits as it's not the goal for this crate currently.
+
+use crate::{
+    errors::{self, Error, ErrorKind},
+    Algorithm,
+};
 use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
-use std::fmt;
+use std::{fmt, str::FromStr};
 
 /// The intended usage of the public `KeyType`. This enum is serialized `untagged`
 #[derive(Clone, Debug, Eq, PartialEq, Hash)]
@@ -142,6 +145,87 @@ impl<'de> Deserialize<'de> for KeyOperations {
     }
 }
 
+/// The algorithms of the keys
+#[allow(non_camel_case_types, clippy::upper_case_acronyms)]
+#[derive(Debug, PartialEq, Eq, Hash, Copy, Clone, Serialize, Deserialize)]
+pub enum KeyAlgorithm {
+    /// HMAC using SHA-256
+    HS256,
+    /// HMAC using SHA-384
+    HS384,
+    /// HMAC using SHA-512
+    HS512,
+
+    /// ECDSA using SHA-256
+    ES256,
+    /// ECDSA using SHA-384
+    ES384,
+
+    /// RSASSA-PKCS1-v1_5 using SHA-256
+    RS256,
+    /// RSASSA-PKCS1-v1_5 using SHA-384
+    RS384,
+    /// RSASSA-PKCS1-v1_5 using SHA-512
+    RS512,
+
+    /// RSASSA-PSS using SHA-256
+    PS256,
+    /// RSASSA-PSS using SHA-384
+    PS384,
+    /// RSASSA-PSS using SHA-512
+    PS512,
+
+    /// Edwards-curve Digital Signature Algorithm (EdDSA)
+    EdDSA,
+
+    /// RSAES-PKCS1-V1_5
+    RSA1_5,
+
+    /// RSAES-OAEP using SHA-1
+    #[serde(rename = "RSA-OAEP")]
+    RSA_OAEP,
+
+    /// RSAES-OAEP-256 using SHA-2
+    #[serde(rename = "RSA-OAEP-256")]
+    RSA_OAEP_256,
+}
+
+impl FromStr for KeyAlgorithm {
+    type Err = Error;
+    fn from_str(s: &str) -> errors::Result<Self> {
+        match s {
+            "HS256" => Ok(KeyAlgorithm::HS256),
+            "HS384" => Ok(KeyAlgorithm::HS384),
+            "HS512" => Ok(KeyAlgorithm::HS512),
+            "ES256" => Ok(KeyAlgorithm::ES256),
+            "ES384" => Ok(KeyAlgorithm::ES384),
+            "RS256" => Ok(KeyAlgorithm::RS256),
+            "RS384" => Ok(KeyAlgorithm::RS384),
+            "PS256" => Ok(KeyAlgorithm::PS256),
+            "PS384" => Ok(KeyAlgorithm::PS384),
+            "PS512" => Ok(KeyAlgorithm::PS512),
+            "RS512" => Ok(KeyAlgorithm::RS512),
+            "EdDSA" => Ok(KeyAlgorithm::EdDSA),
+            "RSA1_5" => Ok(KeyAlgorithm::RSA1_5),
+            "RSA-OAEP" => Ok(KeyAlgorithm::RSA_OAEP),
+            "RSA-OAEP-256" => Ok(KeyAlgorithm::RSA_OAEP_256),
+            _ => Err(ErrorKind::InvalidAlgorithmName.into()),
+        }
+    }
+}
+
+impl fmt::Display for KeyAlgorithm {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
+
+impl KeyAlgorithm {
+    fn to_algorithm(self) -> errors::Result<Algorithm> {
+        Algorithm::from_str(self.to_string().as_str())
+    }
+}
+
 /// Common JWK parameters
 #[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize, Default, Hash)]
 pub struct CommonParameters {
@@ -159,15 +243,15 @@ pub struct CommonParameters {
     #[serde(rename = "key_ops", skip_serializing_if = "Option::is_none", default)]
     pub key_operations: Option<Vec<KeyOperations>>,
 
-    /// The algorithm intended for use with the key
+    /// The algorithm keys intended for use with the key.
     #[serde(rename = "alg", skip_serializing_if = "Option::is_none", default)]
-    pub algorithm: Option<Algorithm>,
+    pub key_algorithm: Option<KeyAlgorithm>,
 
     /// The case sensitive Key ID for the key
     #[serde(rename = "kid", skip_serializing_if = "Option::is_none", default)]
     pub key_id: Option<String>,
 
-    /// X.509 Public key cerfificate URL. This is currently not implemented (correctly).
+    /// X.509 Public key certificate URL. This is currently not implemented (correctly).
     ///
     /// Serialized to `x5u`.
     #[serde(rename = "x5u", skip_serializing_if = "Option::is_none")]
@@ -194,24 +278,20 @@ pub struct CommonParameters {
 
 /// Key type value for an Elliptic Curve Key.
 /// This single value enum is a workaround for Rust not supporting associated constants.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize, Deserialize, Hash)]
+#[derive(Clone, Copy, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Hash)]
 pub enum EllipticCurveKeyType {
     /// Key type value for an Elliptic Curve Key.
+    #[default]
     EC,
-}
-
-impl Default for EllipticCurveKeyType {
-    fn default() -> Self {
-        EllipticCurveKeyType::EC
-    }
 }
 
 /// Type of cryptographic curve used by a key. This is defined in
 /// [RFC 7518 #7.6](https://tools.ietf.org/html/rfc7518#section-7.6)
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, Hash)]
+#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Hash)]
 pub enum EllipticCurve {
     /// P-256 curve
     #[serde(rename = "P-256")]
+    #[default]
     P256,
     /// P-384 curve
     #[serde(rename = "P-384")]
@@ -222,12 +302,6 @@ pub enum EllipticCurve {
     /// Ed25519 curve
     #[serde(rename = "Ed25519")]
     Ed25519,
-}
-
-impl Default for EllipticCurve {
-    fn default() -> Self {
-        EllipticCurve::P256
-    }
 }
 
 /// Parameters for an Elliptic Curve Key
@@ -250,16 +324,11 @@ pub struct EllipticCurveKeyParameters {
 
 /// Key type value for an RSA Key.
 /// This single value enum is a workaround for Rust not supporting associated constants.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize, Deserialize, Hash)]
+#[derive(Clone, Copy, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Hash)]
 pub enum RSAKeyType {
     /// Key type value for an RSA Key.
+    #[default]
     RSA,
-}
-
-impl Default for RSAKeyType {
-    fn default() -> Self {
-        RSAKeyType::RSA
-    }
 }
 
 /// Parameters for a RSA Key
@@ -280,17 +349,12 @@ pub struct RSAKeyParameters {
 
 /// Key type value for an Octet symmetric key.
 /// This single value enum is a workaround for Rust not supporting associated constants.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize, Deserialize, Hash)]
+#[derive(Clone, Copy, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Hash)]
 pub enum OctetKeyType {
     /// Key type value for an Octet symmetric key.
     #[serde(rename = "oct")]
+    #[default]
     Octet,
-}
-
-impl Default for OctetKeyType {
-    fn default() -> Self {
-        OctetKeyType::Octet
-    }
 }
 
 /// Parameters for an Octet Key
@@ -306,17 +370,12 @@ pub struct OctetKeyParameters {
 
 /// Key type value for an Octet Key Pair.
 /// This single value enum is a workaround for Rust not supporting associated constants.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize, Deserialize, Hash)]
+#[derive(Clone, Copy, Debug, Default, Eq, PartialEq, Serialize, Deserialize, Hash)]
 pub enum OctetKeyPairType {
     /// Key type value for an Octet Key Pair.
     #[serde(rename = "OKP")]
+    #[default]
     OctetKeyPair,
-}
-
-impl Default for OctetKeyPairType {
-    fn default() -> Self {
-        OctetKeyPairType::OctetKeyPair
-    }
 }
 
 /// Parameters for an Octet Key Pair
@@ -352,6 +411,13 @@ pub struct Jwk {
     pub algorithm: AlgorithmParameters,
 }
 
+impl Jwk {
+    /// Find whether the Algorithm is implemented and supported
+    pub fn is_supported(&self) -> bool {
+        self.common.key_algorithm.unwrap().to_algorithm().is_ok()
+    }
+}
+
 /// A JWK set
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct JwkSet {
@@ -373,8 +439,10 @@ mod tests {
     use crate::serialization::b64_encode;
     use crate::Algorithm;
     use serde_json::json;
+    use wasm_bindgen_test::wasm_bindgen_test;
 
     #[test]
+    #[wasm_bindgen_test]
     fn check_hs256() {
         let key = b64_encode("abcdefghijklmnopqrstuvwxyz012345");
         let jwks_json = json!({
@@ -392,7 +460,9 @@ mod tests {
         assert_eq!(set.keys.len(), 1);
         let key = &set.keys[0];
         assert_eq!(key.common.key_id, Some("abc123".to_string()));
-        assert_eq!(key.common.algorithm, Some(Algorithm::HS256));
+        let algorithm = key.common.key_algorithm.unwrap().to_algorithm().unwrap();
+        assert_eq!(algorithm, Algorithm::HS256);
+
         match &key.algorithm {
             AlgorithmParameters::OctetKey(key) => {
                 assert_eq!(key.key_type, OctetKeyType::Octet);