+++ tests/eddsa/mod.rs
@@ -3,6 +3,7 @@ use jsonwebtoken::{
     Algorithm, DecodingKey, EncodingKey,
 };
 use serde::{Deserialize, Serialize};
+use wasm_bindgen_test::wasm_bindgen_test;
 
 #[cfg(feature = "use_pem")]
 use jsonwebtoken::{decode, encode, Header, Validation};
@@ -17,6 +18,7 @@ pub struct Claims {
 }
 
 #[test]
+#[wasm_bindgen_test]
 fn round_trip_sign_verification_pk8() {
     let privkey = include_bytes!("private_ed25519_key.pk8");
     let pubkey = include_bytes!("public_ed25519_key.pk8");
@@ -31,6 +33,7 @@ fn round_trip_sign_verification_pk8() {
 
 #[cfg(feature = "use_pem")]
 #[test]
+#[wasm_bindgen_test]
 fn round_trip_sign_verification_pem() {
     let privkey_pem = include_bytes!("private_ed25519_key.pem");
     let pubkey_pem = include_bytes!("public_ed25519_key.pem");
@@ -49,6 +52,7 @@ fn round_trip_sign_verification_pem() {
 
 #[cfg(feature = "use_pem")]
 #[test]
+#[wasm_bindgen_test]
 fn round_trip_claim() {
     let privkey_pem = include_bytes!("private_ed25519_key.pem");
     let pubkey_pem = include_bytes!("public_ed25519_key.pem");
@@ -74,6 +78,7 @@ fn round_trip_claim() {
 
 #[cfg(feature = "use_pem")]
 #[test]
+#[wasm_bindgen_test]
 fn ed_x() {
     let privkey = include_str!("private_ed25519_key.pem");
     let my_claims = Claims {
@@ -99,6 +104,7 @@ fn ed_x() {
 
 #[cfg(feature = "use_pem")]
 #[test]
+#[wasm_bindgen_test]
 fn ed_jwk() {
     use jsonwebtoken::jwk::Jwk;
     use serde_json::json;